Remove noex::TableAdmin.
Remove the class and the tests associated with it. There is probably
some more dead code related to async functions, but we have open bugs to
delete that code too.
