+++ google/cloud/bigtable/bigtable_client.bzl
@@ -62,7 +62,6 @@ bigtable_client_hdrs = [
     "internal/rowreaderiterator.h",
     "internal/strong_type.h",
     "internal/table.h",
-    "internal/table_admin.h",
     "internal/unary_client_utils.h",
     "idempotent_mutation_policy.h",
     "mutation_batcher.h",
@@ -105,7 +104,6 @@ bigtable_client_srcs = [
     "internal/readrowsparser.cc",
     "internal/rowreaderiterator.cc",
     "internal/table.cc",
-    "internal/table_admin.cc",
     "idempotent_mutation_policy.cc",
     "mutation_batcher.cc",
     "mutations.cc",
