+++ google/cloud/bigtable/bigtable_client_unit_tests.bzl
@@ -37,7 +37,6 @@ bigtable_client_unit_tests = [
     "instance_admin_test.cc",
     "instance_config_test.cc",
     "instance_update_config_test.cc",
-    "internal/async_check_consistency_test.cc",
     "internal/async_longrunning_op_test.cc",
     "internal/async_poll_op_test.cc",
     "internal/async_retry_multi_page_test.cc",
@@ -48,7 +47,6 @@ bigtable_client_unit_tests = [
     "internal/table_async_check_and_mutate_row_test.cc",
     "internal/grpc_error_delegate_test.cc",
     "internal/prefix_range_end_test.cc",
-    "internal/table_admin_test.cc",
     "internal/table_async_apply_test.cc",
     "internal/table_async_bulk_apply_test.cc",
     "internal/table_async_row_reader_test.cc",
