+++ google/cloud/bigtable/internal/async_retry_unary_rpc_and_poll_test.cc
@@ -14,7 +14,7 @@
 
 #include "google/cloud/bigtable/internal/async_retry_unary_rpc_and_poll.h"
 #include "google/cloud/bigtable/admin_client.h"
-#include "google/cloud/bigtable/internal/table_admin.h"
+#include "google/cloud/bigtable/completion_queue.h"
 #include "google/cloud/bigtable/testing/internal_table_test_fixture.h"
 #include "google/cloud/bigtable/testing/mock_admin_client.h"
 #include "google/cloud/bigtable/testing/mock_completion_queue.h"
@@ -23,6 +23,7 @@
 #include "google/cloud/bigtable/testing/table_test_fixture.h"
 #include "google/cloud/testing_util/assert_ok.h"
 #include "google/cloud/testing_util/chrono_literals.h"
+#include <google/bigtable/admin/v2/bigtable_table_admin.grpc.pb.h>
 #include <gmock/gmock.h>
 #include <thread>
