+++ google/cloud/bigtable/tests/CMakeLists.txt
@@ -18,7 +18,6 @@
 # Bigtable emulator.
 set(bigtable_client_integration_tests
     admin_integration_test.cc
-    admin_async_integration_test.cc
     admin_async_future_integration_test.cc
     data_async_future_integration_test.cc
     data_async_integration_test.cc
