+++ google/cloud/bigtable/tests/bigtable_client_integration_tests.bzl
@@ -18,7 +18,6 @@
 
 bigtable_client_integration_tests = [
     "admin_integration_test.cc",
-    "admin_async_integration_test.cc",
     "admin_async_future_integration_test.cc",
     "data_async_future_integration_test.cc",
     "data_async_integration_test.cc",
